+++ packages/pieces/mailchimp/src/index.ts
@@ -1,5 +1,5 @@
 import { PieceAuth, createPiece } from '@activepieces/pieces-framework';
-import { addMemberToList } from './lib/actions/add-member-to-list.action/add-member-to-list.action'
+import { addMemberToList } from './lib/actions/add-member-to-list.action';
 import { mailChimpSubscribeTrigger } from './lib/triggers/subscribe-trigger';
 
 export const mailchimpAuth = PieceAuth.OAuth2({
